requirements.txt
requests==2.32.*
beautifulsoup4==4.12.*
lxml==5.*
playwright==1.44.*
PyYAML==6.*
Jinja2==3.1.*
python-dotenv==1.0.*
